Move all user channel tests into one file
