--- functional_tests/test_multiple_users_multiple_channels.py
@@ -1,60 +1,0 @@
-import time
-
-import pytest
-
-
-from lib.admin import Admin
-from lib.verify import verify_changes
-
-from fixtures import cluster
-
-
-@pytest.mark.distributed_index
-@pytest.mark.sanity
-def test_multiple_users_multiple_channels(cluster):
-
-    cluster.reset(config="sync_gateway_default_functional_tests.json")
-
-    # TODO Parametrize
-    num_docs_seth = 10
-    num_docs_adam = 20
-    num_docs_traun = 30
-
-    sgs = cluster.sync_gateways
-
-    admin = Admin(sgs[0])
-
-    seth = admin.register_user(target=sgs[0], db="db", name="seth", password="password", channels=["ABC"])
-    adam = admin.register_user(target=sgs[0], db="db", name="adam", password="password", channels=["NBC", "CBS"])
-    traun = admin.register_user(target=sgs[0], db="db", name="traun", password="password", channels=["ABC", "NBC", "CBS"])
-
-    # TODO use bulk docs
-    seth.add_docs(num_docs_seth)  # ABC
-    adam.add_docs(num_docs_adam)  # NBC, CBS
-    traun.add_docs(num_docs_traun)  # ABC, NBC, CBS
-
-    assert len(seth.cache) == num_docs_seth
-    assert len(adam.cache) == num_docs_adam
-    assert len(traun.cache) == num_docs_traun
-
-    # discuss appropriate time with team
-    time.sleep(10)
-
-    # Seth should get docs from seth + traun
-    seth_subset = [seth.cache, traun.cache]
-    seth_expected_docs = {k: v for cache in seth_subset for k, v in cache.items()}
-    verify_changes([seth], expected_num_docs=num_docs_seth + num_docs_traun, expected_num_updates=0, expected_docs=seth_expected_docs)
-
-    # Adam should get docs from adam + traun
-    adam_subset = [adam.cache, traun.cache]
-    adam_expected_docs = {k: v for cache in adam_subset for k, v in cache.items()}
-    verify_changes([adam], expected_num_docs=num_docs_adam + num_docs_traun, expected_num_updates=0, expected_docs=adam_expected_docs)
-
-    # Traun should get docs from seth + adam + traun
-    traun_subset = [seth.cache, adam.cache, traun.cache]
-    traun_expected_docs = {k: v for cache in traun_subset for k, v in cache.items()}
-    verify_changes([traun], expected_num_docs=num_docs_seth + num_docs_adam + num_docs_traun, expected_num_updates=0, expected_docs=traun_expected_docs)
-
-
-
-
